requirements.txt
Backtesting==0.3.3
beautifulsoup4==4.12.3
bokeh==3.5.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
frozendict==2.4.4
html5lib==1.1
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
multitasking==0.0.11
numpy==2.0.0
packaging==24.1
pandas==2.2.2
peewee==3.17.5
pillow==10.4.0
platformdirs==4.2.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
reportlab==4.2.2
requests==2.32.3
setuptools==70.2.0
six==1.16.0
soupsieve==2.5
ta==0.11.0
tornado==6.4.1
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
xyzservices==2024.6.0
yfinance==0.2.40
